+++ android/history.md
@@ -3,6 +3,9 @@
 ## 13.0 alpha
 * Start version 13.0
 
+## 2019-08-06 12.0.4076 beta
+* Adjustments to Settings UI (#1931)
+
 ## 2019-08-05 12.0.4075 beta
 * Fixes issue with suggestion text misalignment (#1932)